Changes to make it possible to LD_PRELOAD libttif
1) if GST_USING_PRINTF_EXTENSION, then prepend the fmt string with "<%P> " and
pass object as a normal arg.  When using TTIF, you want the whole fmt string,
including the object name prefix, to be constant.  This way, only the fmt
string pointer needs to be logged.
2) GstDebugTraceLocation: small optimization to stash __FILE__, __LINE__, and
GST_FUNCTION together and pass as a single ptr.. the optimization is probably
lost in the noise with the default printf() based traces, but makes more of a
difference with faster trace systems

Incorporate modifications by Ricardo Salveti de Araujo in order to work out
some errors from recent versions of GStreamer.